Organized and added sound resources
