+++ README.md
@@ -1,0 +1,33 @@
+# Galaga clone
+
+Dependencies
+    - Python 3
+    - pygame
+    - elvector
+
+Hardware
+  - screen
+  - input
+  - sound
+  - sound
+
+Input (min)
+  - 4 directions
+  - fire button
+
+TODO
+  - main menu
+    - show scores
+    - title
+    - font
+  - play
+    - stages
+    - enemy programming
+    - player abduction
+      - double ship
+      - captured ship
+      - tractor beam
+    - shooting
+    - show stats on game over
+  - high score entry
+    - interface
