+++ main.py
@@ -1,0 +1,14 @@
+#!/usr/bin/env python3
+
+__author__ = "Izak Halseide"
+
+import sys
+
+import pygame
+
+from source import main
+
+if __name__ == '__main__':
+    main.main()
+    pygame.quit()
+    sys.exit()
