+++ resources/audio/sound index.txt
@@ -1,0 +1,23 @@
+Seconds = Sound effect
+0-7 = Echo theme
+7.64 - 13.56 = Victory thing
+14 - 20 = High score song
+21 - 36 = Game over song
+36 - 40 = Tractor beam
+40 - 48 = Void bloop song
+48.57 - 49.24 = Beginning toot
+50.59 - 53.06 = Whirl whoop
+53.40 - 59.69 = Start song
+60.24 - 66.65 = abducted song echo
+67 - 69.31 = Quick abducted song
+69.64 - 70.72 = Victory toot
+72.12 - 74.61 = Player explosion detailed
+75 - 75.32 = Enemy destroyed
+76.06 - 76.62 = Enemy toots
+77.57 - 78 = Enemy destroyed 2
+78.92 - 79.57 = Enemy destroy streak
+80.76 - 82.61 = Enemy missile
+83.19 - 83.74 = Tri toot
+84.19 - 84.46 = Blooeep
+85.03 - 85.38 = Player fire
+86.22 - 86.36 = Stage beep
