+++ source/constants.py
@@ -1,0 +1,13 @@
+
+# Game States
+MENU_STATE = "Start"
+PLAY_STATE = "Play"
+SCORED_STATE = "Scored"
+
+INITIAL_STATE = PLAY_STATE
+
+WIDTH, HEIGHT = 224, 288
+SCREEN_SIZE = (WIDTH, HEIGHT)
+CENTER_X, CENTER_Y = WIDTH//2, HEIGHT//2
+SCREEN_CENTER = (CENTER_X, CENTER_Y)
+FPS = 30
